+++ koda_validate/_validate_and_map.py
@@ -5,7 +5,7 @@ from typing import Callable, Optional, Tuple, TypeVar, Union, cast, overload
 
 from koda import Err, Ok, Result
 
-from koda_validate.utils import _flat_map_same_type_if_not_none
+from koda_validate._internals import _flat_map_same_type_if_not_none
 
 T1 = TypeVar("T1")
 T2 = TypeVar("T2")
