+++ koda_validate/none.py
@@ -3,8 +3,8 @@ from typing import Any, Final, Optional
 from koda import Err, Ok, Result
 from koda._generics import A
 
+from koda_validate._internals import _variant_errors
 from koda_validate.typedefs import Serializable, Validator
-from koda_validate.utils import _variant_errors
 
 OK_NONE: Final[Ok[None]] = Ok(None)
 OK_NONE_OPTIONAL: Final[Ok[Optional[Any]]] = Ok(None)
